cass/cass.pro changed to allow for so-libraries in LCLS_software:
under linux the LCLS-library had been created from the tarball
using "gmake i386-linux" (also the architecture x86_64-linux is allowed).
These libraries are not part of the "diode" repository.

To compile/run the program Qt, root and LCLS softwares need
to be installed. Since for the time being the LCLS software 
is not platform independent we need to find a solution.

TODO:
1) remove LCLS header files from repository,
2) change pro-files to take care of these changes