implemented pp102.
We need to make sure that the dependencys are build correctly. The
dependency for this postprocessor will change, when the user sets
different parameters. The dependency will be created when creating this
pp. But then it will already check whether the dependend histgrams are
present. If they are not, then we cannot allocte the histogram for it.
In this case, we need to create the other pp first and then rerun
loadsettings for this postprocessor.

This means we need to catch the exeption that is thrown when a pp is not
valid.