make sure that the condition is called only when its initialized.
when condition is not initialized or the condition is true then we
should cass process to process the event. Otherwise we should not call
it, but just put the event at the second to first position. This makes
sure that when someone tries to retrieve a processed histogram it will
always be the front one.