added a postprocessor (pp 66) that will take two 1D histograms and create a 2D histogram from it. This can be used for covariance maps.
When filling histograms make the resulting histogram display how many
times it has been filled. The default value of 0 will cause the
lucassviewer to not display anything.

Implemented the histogramChanged functions for postprocessors that need
it. Corrected the implementation of this function of postprocessors that
already had this implemented.

Give the history pp (64) a default value for the axis.

Debugged the subset histogram pp. This includes correcting the logic and
implementing the histogramChanged function which will update the
histogram and subset values according to the changed dependance