debugged the operator, now it makes sure that when calling the process implemention or the dependands operator that the histogram list will be in a defined state. Calling either function could cause the histogram list to be changed, because the either the own histogram will be resized or a dependand will resize and call histogramsChanged function.
debugged the createHistList function. It will now create the histlist,
so that it looks like it does when the operator will call process for
that pp. This means that the pointer to the current result is contained
in the last element of the histogram list.

todo debug the setupDependency function (started)