some more layout changes
put licence back into the mainpage to make it more prominent.

removed the first section things, because the title of the
page will be displayed anyway. The title of the page is the
same as the title of the first section anyway.

replaces sub- and subsubsections by sections because the html
output will does not differenciate between something just
bold and beiing a sub- or subsubsection.