implemented a new way of setting up the format converters.
Now when one wants to add a new converter one has just to add it in the
backend. The converter nows for which ids it is responsible for.

The user has to give activate the converter by giving a list of the
converters.

Resource managment is now done with a shared pointer. To make it more
exeption safe.