made the acqiris converters singletons
made the commercial ccd converter a singleton

made machine converter a singleton

made the pnccd converter a singleton

moved the blank converter to own namespace and made it a singleton

make use of the singletons

since we are now using the singletons, we don't need a container with
all available converters anymore. Also we don't need a list of keys for
these converters.