reordered startup and added try catch mechanism.
reordered the way the workers and the input is started to make sure that
when creating the workers fails the shared memory message buffers are
not affected.

put the the start of the threads and the qt event loop into a try-catch
statement so that when an exception is thrown we make sure that the
shared memory input will free all message buffers before closing the
program.