make http response functions member functions of http_server class. Add static member function for callback. Make use of callback additional parameter that stores http_server instance and route the callback to the corresponding instance member functions.
http_server now holds a reference to the histogram-getter.

the response classes hold a reference to the http_server, from
which they can get additional information about the request.

a 2dImage response type now returns the jpeg image of the histogram
that is requested. (therefore the address is parsed for the histogram
name).

the 2dPage embeds the 2dImage.
Auto-refresh! (hardcoded to 1 second)

parseAddress hands http_server instance to the response objects it creates.