some bugfixes and extensions
made the recursion depth a user choosable variable

added a check for the size of the incomming pixellist and let the user
choose the maximum size that will still be treated in this function

added information about how many pixels where involved in the hit to the
hit.

bugfixed the check (was not checking for 0 before, now it is checking
for value to be 0). Therefore also added some commented output