changed the way that file input and multi file input works.
the file readers have now a member that will read the header information
from a file.

the map that maps events to file positions is now made of a vector of
pairs of filepointers and their associated readers.

the parser now has a member that contains the filepointer together with
the associated filereader and will save this in the map.

multifile input creates now a reader for every file and reads the header
of the file before the reader together with the file is used to create
the fileparser. This enables us to not read for each event the
fileheader of the file containing the event first.

the parser do not check wether the event is valid anymore. This is done
in the savepos member of the base class. to know what type of parser had
a invalid event id each file parser has now a member returning what kind
it reads.