added header information for raw sss files and make use of it
ensure that the readers will return the eventid to indicate a bad id.
took out the eventid check from the readers since it is not necessary,
since it will be checked before in the parser for multifileinput and in
the fileinput itselve as it checks the return value of the operator.