moved setting of the QMAKE_CLEAN variables into the project files themselves.
Otherwise the main cass.pro project file will also create cleaning
targets in the Makefile, which will cause the Makefile trying to delete
the folder cass.
Also in cass_bin.pro project file the target name is changed and would
not be deleted when the "official" target called cass_bin is put to the
delete list.