change throwing error when timeout occurs to a return value
When a timeout occurs when retrieving and element to process or to fill from the ringbuffer, return a
iterator to the end of the buffer instead of throwing an exception. This allows easier code to periodically
check the status if timeouts are expected. Rewrote the worker to use the new way of telling about a timeout
occurance.